+++ src/app/shared/models/filters.ts
@@ -1,6 +1,0 @@
-export enum FilterType {
-  PLOT = 'plot',
-  YEAR = 'y',
-  TITLE = 't',
-  SEARCH = 's',
-}