+++ src/environments/environment.ts
@@ -1,5 +1,5 @@
 export const environment = {
   production: true,
-  version: '0.0.2',
+  version: '0.0.3',
   BASE_URL: 'https://www.omdbapi.com/?apikey=df30ee18',
 };