+++ changelog.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Glossary](#glossary)
 
+## [0.0.3] - 2024-01-05
+
+### Added
+
+- Custom filters component
+- Search favorites
+
 ## [0.0.2] - 2024-01-05
 
 ### Added